feat(discover): server-render a crawlable content strip for SEO
/discover was a pure client app, so its listings were invisible to search
engines and AI crawlers (the master plan's "Discover ships zero HTML"). The
page is now a server component that fetches a small "recently published" set
server-side and slots it into the interactive client page as server-rendered
HTML — real content + deep links to the already-SSR'd entity detail pages.

- discoverServerFetch.ts: admin-client fetch (no cookies → ISR-cacheable; same
  public-status-filter pattern as sitemap.ts), latest few per high-signal type,
  merged newest-first, capped at 12. Never throws → [] on failure.
- DiscoverFeatured.tsx: server-rendered <section> of entity links (null when empty).
- page.tsx: server entry, revalidate=300, passes the strip as `topContent`.
- Former page.tsx → DiscoverPageClient.tsx (unchanged logic) accepting `topContent`.

Scope: this is the proportionate SEO fix. The primary crawl mechanism — a
sitemap enumerating every public entity URL + SSR'd entity detail pages — is
already in place; full SSR of the interactive result grid would mean seeding 6
nested client hooks (incl. fragile empty-state/loading logic) for marginal gain.
Route is now ○ (static + ISR) instead of a client shell; type-check + lint clean;
production build green.
